--- src/main/kotlin/com/widehouse/cafe/cafe/data/CafeQueryRepository.kt
@@ -1,5 +1,0 @@
-package com.widehouse.cafe.cafe.data
-
-interface CafeQueryRepository {
-    fun findAllByCategoryId(categoryId: Long): List<Cafe>
-}
