--- src/main/kotlin/com/widehouse/cafe/cafe/data/CategoryQueryRepository.kt
@@ -1,5 +1,0 @@
-package com.widehouse.cafe.cafe.data
-
-interface CategoryQueryRepository {
-    fun findByName(name: String): Category?
-}
