+++ src/main/kotlin/com/widehouse/cafe/cafe/repository/CafeRepository.kt
@@ -1,5 +1,6 @@
-package com.widehouse.cafe.cafe.data
+package com.widehouse.cafe.cafe.repository
 
+import com.widehouse.cafe.cafe.model.Cafe
 import org.springframework.data.querydsl.QuerydslPredicateExecutor
 import org.springframework.data.repository.CrudRepository
