+++ src/main/kotlin/com/widehouse/cafe/cafe/repository/CategoryQueryRepository.kt
@@ -1,0 +1,7 @@
+package com.widehouse.cafe.cafe.repository
+
+import com.widehouse.cafe.cafe.model.Category
+
+interface CategoryQueryRepository {
+    fun findByName(name: String): Category?
+}
