+++ src/main/kotlin/com/widehouse/cafe/cafe/repository/CategoryRepository.kt
@@ -1,5 +1,6 @@
-package com.widehouse.cafe.cafe.data
+package com.widehouse.cafe.cafe.repository
 
+import com.widehouse.cafe.cafe.model.Category
 import org.springframework.data.repository.CrudRepository
 
 interface CategoryRepository : CrudRepository<Category, Long>, CategoryQueryRepository
