+++ src/test/kotlin/com/widehouse/cafe/cafe/repository/CafeRepositoryTest.kt
@@ -1,5 +1,8 @@
-package com.widehouse.cafe.cafe.data
+package com.widehouse.cafe.cafe.repository
 
+import com.widehouse.cafe.cafe.model.Cafe
+import com.widehouse.cafe.cafe.model.Category
+import com.widehouse.cafe.cafe.model.QCafe
 import com.widehouse.cafe.config.QueryDslConfig
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.extensions.spring.SpringExtension
